添加了模糊搜索 完善了删除和创建活动
查看某个社团的活动没有数据是因为数据太少，118号社团就有记录

activity部分还有两个问题，
1.创建新活动因为没有输入活动id，所以不知道活动id，没法和club表连接起来
2.获取活动介绍都为null，不知道为什么

+++ After_End/src/view/ActivityView.java
@@ -24,13 +24,14 @@ public class ActivityView extends View {
                         + "3.创建一个活动\n"
                         + "4.查看社团一年的活动\n"
                         + "5.删除活动\n"
-                        + "6.退出\n"
+                        + "6.模糊查询\n"
+                        + "7.退出\n"
         );
     }
 
     public void displayMenu() {
         int instruction;
-        final int EXIT = 6;
+        final int EXIT = 7;
 
         while (true) {
             display();
@@ -62,6 +63,12 @@ public class ActivityView extends View {
                         //todo 删除活动,只能删自己创建的活动
                         deleteMyActivity();
                         break;
+                    case 6:
+                        //todo 模糊查询
+                        System.out.println("请输入关键词");
+                        String word = in.next();
+                        activitySearch(word);
+                        break;
                     default:
                         throw new CustomerException("Wrong input");
                 }
@@ -75,51 +82,22 @@ public class ActivityView extends View {
         //todo activityBrowse 最近一个个月的活动
         activityService.showActivityMonth();
         long num = activityService.numActivityMonth();
-
-        String prompt = "----------\n" +
-                "最近一月没有活动\n" +
-                "----------\n";
-        String content = "1.选择查看详细信息\n" +
-                "2.返回\n>";
-        int instruction = -1;
-        final int EXIT = 2;
-        while (true) {
-            //todo 获取所有活动
-            if (num == 0)//如果没有活动
-                System.out.print(prompt);
-            System.out.print(content);
-
-            try {
-                instruction = InputInt(in);
-
-                if (instruction == EXIT)
-                    break;
-
-                switch (instruction) {
-                    case 1:
-                        //todo 选择查看活动的详细信息
-                        System.out.println("请输入活动ID");
-                        int ID = InputInt(in);
-                        activityService.showActivityByID(ID);
-                        break;
-                    default:
-                        throw new CustomerException("输入超出范围");
-                }
-            } catch (Exception e) {
-                //todo 异常处理
-            }
-        }
+        detail(num);
     }
-
+    private void activitySearch(String word) throws SQLException{
+        activityService.showActivityword(word);
+        long num = activityService.numActivityByword(word);
+        detail(num);
+    }
     private void activityJoin() throws SQLException {
         activityService.showActivityList(user.getUser_ID());
     }
 
     private void createActivity() throws SQLException {
         System.out.println("请输入活动名称");
-        String name = in.nextLine();
+        String name = in.next();
         System.out.println("请输入活动内容");
-        String content = in.nextLine();
+        String content = in.next();
         System.out.println("请输入活动开始时间\n" + "年：");
         int year = InputInt(in);
         System.out.println("请输入活动开始时间\n" + "月：");
@@ -148,6 +126,7 @@ public class ActivityView extends View {
         int r = InputInt(in);
         boolean range;
         range = (r == 0);
+
         Activity a = new Activity(name, content, start, end, Response_ID, range, true);
         activityService.createActivity(a);
     }
@@ -156,7 +135,10 @@ public class ActivityView extends View {
         //todo activityBrowse 最近一年的活动
         activityService.showActivityYearByClub(id);
         long num = activityService.numActivityByClub(id);
-
+        detail(num);
+    }
+    // 详细查询
+    private void detail(long num){
         String prompt = "----------\n" +
                 "此社团没有活动\n" +
                 "----------\n";
@@ -195,6 +177,11 @@ public class ActivityView extends View {
     private void deleteMyActivity() throws SQLException {
         System.out.println("请输入要删除的活动id");
         int aid = InputInt(in);
-        activityService.deleteActivityByClub(aid, user.getUser_ID());
+        int respon = activityService.activityResponse(aid);
+        if(user.getUser_ID() == respon)
+            activityService.deleteActivityByClub(aid, user.getUser_ID());
+        else{
+            System.out.println("您无权删除此活动");
+        }
     }
 }
